[DFS1] Try to expose raycast to script
[git-p4: depot-paths = "//depot/C35/Students/shangliz/SD/Engine/": change = 465276]

+++ Code/MingEngine/Core/Object/Object.hpp
@@ -10,7 +10,7 @@
 	ClassDatabase::BindConstant(#className, #enumName, static_cast<int>(className::enumName))
 
 #define BIND_CONSTANT(namespaceName, constantName)                                                                     \
-	ClassDatabase::BindConstant(#namespaceName, #constantName, namespaceName::constantName)
+	ClassDatabase::BindConstant(#namespaceName, #constantName, static_cast<int>(namespaceName::constantName))
 
 #define MCLASS(className, inheritName)                                                                                 \
 public:                                                                                                                \
